requirements.txt
openpyxl==3.1.2
pandas==2.2.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.128.0
requests==2.32.3
pyinstaller==6.6.0
